backend/requirements.txt
functions-framework==3.*
flask==2.*
flask-cors==4.*
google-cloud-firestore==2.*
requests==2.*
google-api-core==2.*
pytz==2024.1